requirements.txt
Django==1.11
Markdown==2.6.8
Pillow==4.1.1
argparse==1.2.1
django-filter==1.0.2
djangorestframework==3.6.2
olefile==0.44
pytz==2017.2
wsgiref==0.1.2
